# Add vectorized plotting functionality to improve performance