--- a/manim/mobject/graphing/coordinate_systems.py
+++ b/manim/mobject/graphing/coordinate_systems.py
@@ -1365,7 +1365,7 @@ class CoordinateSystem:
         use_vectorized
             Whether to use the vectorized version of the antiderivative. This means
             to pass in the generated t value array to the function. Only use this if your function supports it.
-            Output should be a numpy array of shape ``[y_0,y_1,...]``
+            Output should be a numpy array of shape ``[y_0, y_1, ...]``
         kwargs
             Any valid keyword argument of :class:`~.ParametricFunction`.