--- a/tests/test_graphical_units/conftest.py
+++ b/tests/test_graphical_units/conftest.py
@@ -6,3 +6,8 @@ import pytest
 @pytest.fixture
 def show_diff(request):
     return request.config.getoption("show_diff")
+
+
+@pytest.fixture(params=[True, False])
+def use_vectorized(request):
+    yield request.param