… manim/mobject/graphing/number_line.py
… manim/mobject/graphing/scale.py
… manim/scene/scene.py
… manim/utils/docbuild/__init__.py
… manim/utils/space_ops.py
… manim/utils/testing/frames_comparison.py
… tests/test_coordinate_system.py
… tests/test_graphical_units/control_data/coordinate_system/plot_log_x_axis_vectorized.npz
… tests/test_graphical_units/control_data/plot/get_riemann_rectangles_use_vectorized[False].npz
… tests/test_graphical_units/control_data/plot/get_riemann_rectangles_use_vectorized[True].npz
… tests/test_graphical_units/control_data/plot/plot_derivative_graph_use_vectorized[False].npz
… tests/test_graphical_units/control_data/plot/plot_derivative_graph_use_vectorized[True].npz
… tests/test_graphical_units/control_data/plot/plot_functions_use_vectorized[False].npz
… tests/test_graphical_units/control_data/plot/plot_functions_use_vectorized[True].npz
… tests/test_graphical_units/control_data/plot/plot_use_vectorized[False].npz
… tests/test_graphical_units/control_data/plot/plot_use_vectorized[True].npz
… tests/test_graphical_units/test_axes.py
… tests/test_graphical_units/test_coordinate_systems.py
… tests/test_number_line.py
… tests/test_space_ops.py