<!-- Thank you for contributing to Manim! Learn more about the process in our contributing guidelines: https://docs.manim.community/en/latest/contributing.html -->

## Overview: What does this pull request change?
<!-- If there is more information than the PR title that should be added to our release changelog, add it in the following changelog section. This is optional, but recommended for larger pull requests. -->
<!--changelog-start-->
- added vectorized logic to `coords_to_points` in `Axes`
- added `use_vectorized: bool` parameter to `plot` in `CoordinateSystem`
- added `use_vectorized: bool` parameter to `ParametricFunction`
- added vectorized logic to `generate_points` in `ParametricFunction` depending on `self.use_vectorized`
- added `use_vectorized: bool` parameter to `plot_antiderivative_graph`
- added vectorized logic to `plot_antiderivative_graph`
- added vectorized logic to `number_to_point` in `NumberLine`
- vectorized `inverse_function` in `scale.py`
- vectorized `angle_of_vector` in `space_ops.py`

- changed naming of test data in `frames_comparison.py` to support fixture parameters because ":" is not allowed in a file name
- added `use_vectorized` pytest fixture
- added `use_vectorized` fixture to
    - `test_axes::test_plot_functions`
    - `test_axes::test_plot_derivative_graph`
    - `test_axes::test_plot`
    - `test_axes::test_get_riemann_rectangles`
- added test `test_coordinate_systems::test_plot_log_x_axis_vectorized`
- added test `test_coordinate_system::test_point_to_coords_vectorized`
- added test `test_coordinate_system::test_coords_to_point_vectorized`
- added test `test_space_ops::test_angle_of_vector_vectorized`
- added test `test_number_line::test_number_to_point`
- added test `test_number_line::test_point_to_number`
- added test `test_space_ops::test_angle_of_vector_vectorized`
<!--changelog-end-->


Todo for Future PR
- [ ] CoordinateSystem.plot_implicit_curve
- [ ] CoordinateSystem.plot_polar_graph
- [ ] ImplicitFunction
- [ ] CoordinateSystem.pr2pt

## Motivation and Explanation: Why and how do your changes improve the library?
<!-- Optional for bugfixes, small enhancements, and documentation-related PRs. Otherwise, please give a short reasoning for your changes. -->

Manim does a ton of function calls internally and reducing them step by step might be a good idea. At some point we could make a breaking change and vectorize functions by default to remove branching code versions. The performance increase is not that much but a step in the right direction. I think this vectorized version could also improve VectorField or other Classes that use `coords_to_points` 🤔

| Before | After Non-Vectorized | Vectorized |
|-|-|-|
|![grafik](https://user-images.githubusercontent.com/17803932/170594656-e270dce3-93ac-4606-9424-a41014b939ce.png)|![grafik](https://user-images.githubusercontent.com/17803932/170764462-36afc503-1b73-4bbc-afbd-9c293767a8b1.png)|![grafik](https://user-images.githubusercontent.com/17803932/170766275-fa2324a5-aa4e-4608-b35b-a1c59482e349.png)|
|![grafik](https://user-images.githubusercontent.com/17803932/170594744-852a63ad-cc36-4959-8fea-d7e3ec94e479.png) `coords_to_point` (yellow) on the right taking 1/3 of `generate_points`|![grafik](https://user-images.githubusercontent.com/17803932/170765060-9fa399af-dda7-468e-9325-0d6c056c8feb.png)| ![grafik](https://user-images.githubusercontent.com/17803932/170766433-9629858b-bd68-4e07-9baf-cce9ec45ab90.png) `coords_to_point` on the left taking a lot less|

### Table Calls/Percentage
Function | Before | After Non-Vectorized | Vectorized |
|-|-|-|-|
`number_to_point` | 327 / 6.3 | 407 / 6.5% | 58 / 1.0% |
`coords_to_point` | 486 / 8% | 517 / 8.2% | 68 / 1.3% |
`generate_points` | 1481 / 24.3% | 1508 / 24.0% | 978 / 18.8% |
**Time (s)** | $42.33$ | $43.48$ | $36.39$ |

As we can see the implementations of `number_to_point` and `coords_to_point` get a little bit slower but i tried to reduce the overhead as much as possible in the non vectorized case to not slow down the library overall. But when more and more functions adapt to the vectorized style this will be neglible because we can see that `generate_points` overall gets faster.

Data collected via `vprof` and performance difference is from 42 to 36 seconds.
To view the [profile_data.zip](https://github.com/ManimCommunity/manim/files/8788980/profile_data.zip) use 
```bash
# original main branch
vprof -i reference.json
# this PR
vprof -i vectorized.json
vprof -i not_vectorized.json
```


# Scene used for testing
![Bench_ManimCE_v0 15 2](https://user-images.githubusercontent.com/17803932/170594268-9ef4f0e7-79b5-4e3c-8a91-78aa2a10a9f6.gif)

```py
from manim import *
class Bench(Scene):
    def construct(self):
        ax1 = NumberPlane((-3, 3), (-4, 4))
        track = ValueTracker(0)
        def func(t):
            out =  np.array([np.sin(2 * (t+track.get_value())), np.sin(3 * (t+track.get_value())), 0])
            return out

        vectorized = False
        graph = ax1.plot_parametric_curve(func, color=RED, t_range=np.array([0, 1]), use_vectorized=vectorized)
        graph2 = ax1.plot(lambda x: x**2, use_vectorized=vectorized)
        graph3 = ax1.plot_derivative_graph(graph2, use_vectorized=vectorized)

        graph.add_updater(lambda m: m.become(ax1.plot_parametric_curve(func, color=RED, t_range=np.array([0, 1]), use_vectorized=vectorized)))
        graph2.add_updater(lambda m: m.become(ax1.plot(lambda x: (x**3)*track.get_value()-x/2, use_vectorized=vectorized)))
        graph3.add_updater(lambda m: m.become(ax1.plot_derivative_graph(graph2, use_vectorized=vectorized)))
        rects = ax1.get_riemann_rectangles(graph)
        rects.add_updater(lambda m: m.become(ax1.get_riemann_rectangles(graph)))

        self.add(ax1, graph, graph2, graph3, rects)
        self.play(track.animate.set_value(TAU),run_time=10)
```

## Links to added or changed documentation pages
<!-- Please add links to the affected documentation pages (edit the description after opening the PR). The link to the documentation for your PR is https://manimce--####.org.readthedocs.build/en/####/, where #### represents the PR number. -->


## Further Information and Comments
<!-- If applicable, put further comments for the reviewers here. -->

This is a extended implementation of
- #2513 

<!-- Thank you again for contributing! Do not modify the lines below, they are for reviewers. -->
## Reviewer Checklist
- [ ] The PR title is descriptive enough for the changelog, and the PR is labeled correctly
- [ ] If applicable: newly added non-private functions and classes have a docstring including a short summary and a PARAMETERS section
- [ ] If applicable: newly added functions and classes are tested