@MrDiver 2022-05-27
Maybe it would be a good idea to add a `use_vectorized` option in the config dictionary