@behackl 2022-05-27
> Maybe it would be a good idea to add a `use_vectorized` option in the config dictionary. This would remove a lot of the unnecessary parameter passing and would make extending this to other functions easier!

I'm not sure about this. The respective classes or objects should hold this sort of behavioral information directly, the global config should only do this for classes that the user should normally not access (like the renderer, the camera, the file writer, ...). And what if you'd want to use vectorized functions in some parts, but not in others?

I'm open to further discussion, but there has to be a very clear and thought through proposal if we want to head in this direction. :-)