@MrDiver 2022-05-27
> And what if you'd want to use vectorized functions in some parts, but not in others?
This is not really possible anyway, beacuse then the function all need to contain logic to decide what their input is and if they need to translate an input array to to elements and then go through a for loop and apply itself. The idea behind that is to slowly optimize the code behind the scenes and give the option to use it but not result in breaking changes so that some function are not working correctly anymore.

The only thing that would still be useful is on the high level end of the plot functions to decide manually if someone would like to have `f(x)` or `f([x_0,x_1,...])`

One option would also be to completely vectorize the low level functions in any way or add a vectorized version for all of them. But this results in a lot of code doubling. And when the goal is to optimize in the long term it seems rather counter productive.